feat: reorganize inner wrapper styles, fix error style bug on filled variant
affects: @medly-components/core, @medly-components/theme
